Move attachements from components folder to pages.
