+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "pip-boy-mod-terminal",
   "description": "A special terminal for giving you a bit more control over your Pip-Boy 3000 Mk V!",
-  "version": "1.5.3",
+  "version": "1.5.4",
   "author": "Cody Tolene <codyatolene@gmail.com>",
   "homepage": "https://λ.guru/",
   "private": true,
