+++ src/app/constants/app-version.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = '1.5.3';
+export const APP_VERSION = '1.5.4';
